+++ src/helpers.php
@@ -3,47 +3,71 @@
 use Illuminate\Support\Debug\Dumper;
 use Carbon\Carbon;
 
-if (!function_exists('vd')) {
+if (!function_exists('dump')) {
     /**
      * Dump les valeurs passées en paramètres à l'aide du Dumper de Laravel,
-     * sans terminer le script à la fin.
+     * sans terminer le script à la fin et en retournant le résultat au lieu
+     * de le rendre directement dans le flux.
+     *
+     * @param  mixed
+     * @return string
+     */
+    function dump()
+    {
+        ob_start();
+
+        array_map(function($x) {(new Dumper)->dump($x)."\n";}, func_get_args());
+
+        return ob_end_clean();
+    }
+}
+
+if (!function_exists('vd')) {
+    /**
+     * Idem dump(), mais rend le résultat directement dans le flux comme var_dump().
      *
      * @param  mixed
      * @return void
      */
     function vd()
     {
-        array_map(function($x) { (new Dumper)->dump($x)."\n"; }, func_get_args());
+        array_map(function($x) {(new Dumper)->dump($x)."\n";}, func_get_args());
     }
 }
 
 if (!function_exists('dw')) {
     /**
-     * Dump les valeurs passées en paramètres à l'aide du Dumper de Laravel,
-     * puis écrit le résultat dans le fichier "public/dump.html".
+     * Idem dump(), mais écrit le résultat dans "public/dump.html".
      *
      * @param  mixed
      * @return void
      */
     function dw()
     {
-        ob_start();
+        $dump = call_user_func_array('dump', func_get_args());
 
-        call_user_func_array('vd', func_get_args());
-
-        file_put_contents(public_path().'/dump.html', ob_get_clean(), FILE_APPEND);
+        file_put_contents(public_path('dump.html'), $dump, FILE_APPEND);
     }
 }
+
 
 if (!function_exists('v')) {
     /**
      * Retourne la valeur d'une variable. Si celle-ci n'existe pas (variable non
      * définie ou chaîne vide), une valeur par défaut est retournée à la place.
      *
+     * ---------------------------------------------------------------------
      * ATTENTION : Si la variable n'existe pas, le passage par référence a pour
      * conséquence de la créer (avec valeur NULL), ce qui peut être problématique
      * si $var est une référence à une dimension d'un tableau ou à un attribut
-     * d'objet car toute la profondeur sera ainsi créée.
+     * d'objet car ces éléments seront alors créés...
+     *
+     * Exemple :
+     *   Si $arr = [] et que l'on fait : v($arr['a']['b'])
+     *   alors $arr vaut à présent : ['a' => ['b' => null]]
+     *
+     * Préférer dans ce cas l'utilisation du helper vv().
+     * ---------------------------------------------------------------------
      *
      * @param  mixed               &$var
      * @param  mixed               $default  Valeur à retourner si $var non défini ou chaîne vide
@@ -68,7 +92,7 @@ if (!function_exists('vv')) {
     /**
      * Idem helper v() mais avec un paramètre $key pour accéder aux dimensions
      * d'un tableau ou attributs d'un objet, évitant ainsi la création de toute
-     * la profondeur si la variable n'existe pas.
+     * la profondeur si celle-ci n'existe pas.
      *
      * @param  mixed               &$var
      * @param  string              $key